api/requirements.txt
fastapi
uvicorn
numpy
pydantic
python-multipart
google-genai
python-dotenv
